Move access_token to internal_utils to avoid import circularity
